Удалил XMLhttp.js переместил код оттуда в main.js в carousel.js код обернул в функции, для дальшейшего вызова при onload событии Удалил старый объект пакетов добавил реализацию запроса к JSON файлу через fetch (закомментировал)
