+++ app/index.html
@@ -5,7 +5,6 @@
 		<title>ЦФТ-Главная</title>
 		<link rel="stylesheet" type="text/css" href="css/uikit.css">
 		<link rel="stylesheet" type="text/css" href="css/style.css">
-		<script type="module" src="js/XMLhttp.js"></script>
 	</head>
 	<body>
 		<header class="o-header">
@@ -238,7 +237,6 @@
 			</symbol>
 		</svg>
 		<script type="module" src="js/drawRandomPackages.js"></script>
-		<script type="module" src="js/XMLhttp.js"></script>
 		<script type="module" src="js/carousel.js"></script>
 		<script type="module" src="js/main.js"></script>
 	</body>
